Fix for GCC on tiger
see http://sourceforge.net/mailarchive/forum.php?thread_name=m2sl5p4o2a.fsf%40gmail.com&forum_name=arabica-devel

+++ include/XSLT/impl/xslt_stylesheet.hpp
@@ -148,7 +148,7 @@ public:
   {
     // entirely simple so far
     StackFrame frame(context);
-    LastFrame last(context, nodes.getLength());
+    LastFrame last(context, (size_t)nodes.getLength());
     for(int i = 0, ie = nodes.getLength(); i != ie; ++i)
     {
       context.setPosition(nodes.item(i), i+1);
